fixing readme
