+++ .pre-commit-config.yaml.txt
@@ -1,0 +1,26 @@
+repos:
+#-   repo: https://github.com/pre-commit/pre-commit-hooks
+#    rev: v3.4.0
+#    hooks:
+#    -   id: trailing-whitespace
+#    -   id: end-of-file-fixer
+#-     repo: https://github.com/psf/black
+#      rev: 20.8b1
+#      hooks:
+#          - id: black
+#            args: [ --config=pyproject.toml ]
+#-     repo: https://github.com/hadialqattan/pycln
+#      rev: v0.0.1-beta.3
+#      hooks:
+#          - id: pycln
+#            args: [ --config=pyproject.toml ]
+#-     repo: https://github.com/pycqa/isort
+#      rev: 5.5.4
+#      hooks:
+#          - id: isort
+#            files: "\\.(py)$"
+#            args: [ --settings-path=pyproject.toml ]
+#-     repo: https://github.com/pre-commit/mirrors-mypy
+##      rev: v0.812
+##      hooks:
+##         - id: mypy
