+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     environment:
       POSTGRES_USER: root
       POSTGRES_PASSWORD: root
-      POSTGRES_DB: feedgate
+      POSTGRES_DB: parking-decision
     ports:
       - "5432:5432"
