+++ parking_decision/business_processes/parking_decider_bp.py
@@ -1,8 +1,7 @@
-from typing import Type
 
+from parking_decision.busniess_components.check_plate_bc import CheckPlateBC
 from parking_decision.dals.decision_dal import DecisionDAL
 from parking_decision.helpers.ocr_space_api import OCRSpaceAPi
-from parking_decision.busniess_components.check_plate_bc import CheckPlateBC
 
 # trying new workflow
