+++ parking_decision/db/models/__init__.py
@@ -1,5 +1,4 @@
 from sqlalchemy.orm import declarative_base, scoped_session, sessionmaker
 
-
 DBSession = scoped_session(sessionmaker())
 Base = declarative_base()
