+++ parking_decision/tests/unit/business_processes/__init__.py
@@ -1,4 +1,3 @@
 #!/usr/bin/env python3
 
 # Author: Simon on 21/07/2022
-
