+++ pyproject.tomll
@@ -1,0 +1,13 @@
+[tool.black]
+line-length = 79
+
+[tool.pycln]
+all = true
+
+[tool.isort]
+line_length = 79
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
