+++ requirements.txt
@@ -1,0 +1,82 @@
+anyio==3.6.1
+asn1crypto==1.5.1
+attrs==22.1.0
+certifi==2022.6.15
+cfgv==3.3.1
+charset-normalizer==2.1.1
+click==8.1.3
+distlib==0.3.6
+fastapi==0.79.1
+fastapi-utils==0.2.1
+filelock==3.8.0
+greenlet==1.1.2
+h11==0.13.0
+identify==2.5.5
+idna==3.3
+iniconfig==1.1.1
+nodeenv==1.7.0
+packaging==21.3
+pg8000==1.29.1
+platformdirs==2.5.2
+pluggy==1.0.0
+pre-commit==2.20.0
+py==1.11.0
+pydantic==1.9.2
+pyparsing==3.0.9
+pytest==7.1.3
+python-multipart==0.0.5
+PyYAML==6.0
+requests==2.28.1
+scramp==1.4.1
+six==1.16.0
+sniffio==1.2.0
+SQLAlchemy==1.4.40
+starlette==0.19.1
+toml==0.10.2
+tomli==2.0.1
+typing_extensions==4.3.0
+urllib3==1.26.11
+uvicorn==0.18.2
+virtualenv==20.16.5
+anyio==3.6.1
+asn1crypto==1.5.1
+attrs==22.1.0
+certifi==2022.6.15
+cfgv==3.3.1
+charset-normalizer==2.1.1
+click==8.1.3
+distlib==0.3.6
+fastapi==0.79.1
+fastapi-utils==0.2.1
+filelock==3.8.0
+greenlet==1.1.2
+h11==0.13.0
+identify==2.5.5
+idna==3.3
+iniconfig==1.1.1
+nodeenv==1.7.0
+packaging==21.3
+pg8000==1.29.1
+platformdirs==2.5.2
+pluggy==1.0.0
+pre-commit==2.20.0
+py==1.11.0
+pydantic==1.9.2
+pyparsing==3.0.9
+pytest==7.1.3
+python-multipart==0.0.5
+PyYAML==6.0
+requests==2.28.1
+scramp==1.4.1
+six==1.16.0
+sniffio==1.2.0
+SQLAlchemy==1.4.40
+starlette==0.19.1
+toml==0.10.2
+tomli==2.0.1
+typing_extensions==4.3.0
+urllib3==1.26.11
+uvicorn==0.18.2
+virtualenv==20.16.5
+
+dill~=0.3.5.1
