--- requirments.txt
@@ -1,30 +1,0 @@
-anyio==3.6.1
-asn1crypto==1.5.1
-attrs==22.1.0
-certifi==2022.6.15
-charset-normalizer==2.1.1
-click==8.1.3
-fastapi==0.79.1
-fastapi-utils==0.2.1
-greenlet==1.1.2
-h11==0.13.0
-idna==3.3
-iniconfig==1.1.1
-packaging==21.3
-pg8000==1.29.1
-pluggy==1.0.0
-py==1.11.0
-pydantic==1.9.2
-pyparsing==3.0.9
-pytest==7.1.3
-python-multipart==0.0.5
-requests==2.28.1
-scramp==1.4.1
-six==1.16.0
-sniffio==1.2.0
-SQLAlchemy==1.4.40
-starlette==0.19.1
-tomli==2.0.1
-typing_extensions==4.3.0
-urllib3==1.26.11
-uvicorn==0.18.2
